refactor: implement component composition and useProducts
