+++ src/components/index.ts
@@ -1,11 +1,10 @@
-export * from "./button";
-export * from "./drawer";
-export * from "./error";
-export * from "./layout";
-export * from "./loader";
-export * from "./main";
-export * from "./navbar";
-export * from "./product-card";
-export * from "./products-list";
-export * from "./search-bar";
-export * from "./sort-by";
+export * from './button';
+export * from './drawer';
+export * from './error';
+export * from './layout';
+export * from './loader';
+export * from './navbar';
+export * from './product-card';
+export * from './products-list';
+export * from './search-bar';
+export * from './sort-by';
